fix: remove uncessary code, combine where obvious
