--- protocol/data_processing/samples/posts.json
@@ -1,102 +1,0 @@
-[
-    {
-        "uid": "1",
-        "user_id": "1470086780",
-        "subnet_id": "59",
-        "query": "(dogecoin) until:2024-12-11 since:2024-12-10",
-        "count": 2,
-        "created_at": 1725519123,
-        "tweets": [
-            {
-                "Tweet": {
-                    "ConversationID": "1831582820713562141",
-                    "GIFs": null,
-                    "Hashtags": null,
-                    "HTML": "<a href=\"https://twitter.com/Cyber_Dogecoin\">@Cyber_Dogecoin</a> $300",
-                    "ID": "1831586071164907637",
-                    "InReplyToStatus": null,
-                    "InReplyToStatusID": "1831582820713562141",
-                    "IsQuoted": false,
-                    "IsPin": false,
-                    "IsReply": true,
-                    "IsRetweet": false,
-                    "IsSelfThread": false,
-                    "Likes": 0,
-                    "Name": "Ifeniboh Onyeka Phillips",
-                    "Mentions": [
-                        {
-                            "ID": "1713362828789465088",
-                            "Username": "Cyber_Dogecoin",
-                            "Name": "CyberDoge"
-                        }
-                    ],
-                    "PermanentURL": "https://twitter.com/stylustica/status/1831586071164907637",
-                    "Photos": null,
-                    "Place": null,
-                    "QuotedStatus": null,
-                    "QuotedStatusID": "",
-                    "Replies": 0,
-                    "Retweets": 0,
-                    "RetweetedStatus": null,
-                    "RetweetedStatusID": "",
-                    "Text": "@Cyber_Dogecoin $300",
-                    "Thread": null,
-                    "TimeParsed": "2024-09-05T06:52:03Z",
-                    "Timestamp": 1725519123,
-                    "URLs": null,
-                    "UserID": "1470086780",
-                    "Username": "stylustica",
-                    "Videos": null,
-                    "Views": 0,
-                    "SensitiveContent": false
-                },
-                "Error": null
-            },
-            {
-                "Tweet": {
-                    "ConversationID": "1779236421296923088",
-                    "GIFs": null,
-                    "Hashtags": null,
-                    "HTML": "<a href=\"https://twitter.com/444worldeth\">@444worldeth</a> Plot twist: they're actually secret agents for Dogecoin 🐕🚀",
-                    "ID": "1831585995717845039",
-                    "InReplyToStatus": null,
-                    "InReplyToStatusID": "1779236421296923088",
-                    "IsQuoted": false,
-                    "IsPin": false,
-                    "IsReply": true,
-                    "IsRetweet": false,
-                    "IsSelfThread": false,
-                    "Likes": 0,
-                    "Name": "Mathias Mateos",
-                    "Mentions": [
-                        {
-                            "ID": "2761277700",
-                            "Username": "444worldeth",
-                            "Name": "444"
-                        }
-                    ],
-                    "PermanentURL": "https://twitter.com/0ce5e50e4efd48c/status/1831585995717845039",
-                    "Photos": null,
-                    "Place": null,
-                    "QuotedStatus": null,
-                    "QuotedStatusID": "",
-                    "Replies": 0,
-                    "Retweets": 0,
-                    "RetweetedStatus": null,
-                    "RetweetedStatusID": "",
-                    "Text": "@444worldeth Plot twist: they're actually secret agents for Dogecoin 🐕🚀",
-                    "Thread": null,
-                    "TimeParsed": "2024-09-05T06:51:45Z",
-                    "Timestamp": 1725519105,
-                    "URLs": null,
-                    "UserID": "2953786353",
-                    "Username": "0ce5e50e4efd48c",
-                    "Videos": null,
-                    "Views": 1,
-                    "SensitiveContent": false
-                },
-                "Error": null
-            }
-        ]
-    }
-]
