--- protocol/scoring/__init__.py
@@ -1,4 +1,0 @@
-from .post_scorer import PostScorer
-from .miner_weights import MinerWeights
-
-__all__ = ['PostScorer', 'WeightScorer']
